refactor: spread application crate
